Can no longer drop a non-plain text table or db, added tests that you can drop/create plain text dbs and tables

+++ src/main/java/edu/uob/Parser.java
@@ -64,10 +64,16 @@ public class Parser {
         moveToNextToken();
         if (query.get(count).equalsIgnoreCase("DATABASE")) {
             moveToNextToken();
+            if (!isPlainText(getCurrentToken())) {
+                throw new Exception("Database names must be plain text, received " + getCurrentToken());
+            }
             databaseList.dropDatabase(getCurrentToken());
         }
         if (query.get(count).equalsIgnoreCase("TABLE")) {
             moveToNextToken();
+            if (!isPlainText(getCurrentToken())) {
+                throw new Exception("Table names must be plain text, received " + getCurrentToken());
+            }
             databaseList.getActiveDB().dropTable(getCurrentToken());
         }
     }
